Sync code with QDriverStation

+++ src/Protocols/FRC_2014.cpp
@@ -170,14 +170,14 @@ DS::SocketType FRC_2014::robotSocketType() {
 }
 
 /**
- * Radio is located at 10.XX.YY.1
+ * Radio is located at 10.TE.AM.1
  */
 QString FRC_2014::defaultRadioAddress() {
     return DS::getStaticIP (10, config()->team(), 1);
 }
 
 /**
- * Robot is located at 10.XX.YY.2
+ * Robot is located at 10.TE.AM.2
  */
 QString FRC_2014::defaultRobotAddress() {
     return DS::getStaticIP (10, config()->team(), 2);
@@ -266,8 +266,8 @@ bool FRC_2014::interpretRobotPacket (const QByteArray& data) {
 
     /* Read status echo code and voltage */
     uint opcode  = data.at (0);
-    uint integer = data.at (1);
-    uint decimal = data.at (2);
+    quint8 integer = data.at (1);
+    quint8 decimal = data.at (2);
 
     /* Parse the voltage (which is stored in a strange format) */
     QString voltage;
